Cleanup: Re-arranged the pca_topk folder structure and moved compute benchmark to top level
