+++ README.md
@@ -9,7 +9,7 @@ Follow the instructions in their readme.
 Say you want to find the perplexity for H2O on a llama-2-7b model
 
 ```
-python -u eval_ppl.py --sequence-length 4096 --model-id meta-llama/Llama-2-7b-hf --model-type llama --use-h2o --heavy-ratio 0.1
+python -u evaluate_tasks.py --sequence-length 4096 --model-id meta-llama/Llama-2-7b-hf --model-type llama --use-h2o --heavy-ratio 0.1
 ```
 
 ### With AxoNN's tensor parallelism
@@ -19,7 +19,7 @@ For this you will need to launch the code using mpirun
 
 
 ```
-mpirun -np 2 python -u eval_ppl.py --sequence-length 4096 --model-id meta-llama/Llama-2-13b-hf --model-type llama --use-h2o --heavy-ratio 0.1 --use-axonn
+mpirun -np 2 python -u evaluate_tasks.py --sequence-length 4096 --model-id meta-llama/Llama-2-13b-hf --model-type llama --use-h2o --heavy-ratio 0.1 --use-axonn
 ```
