+++ examples/pca_topk/job_template_perlmutter.sh
@@ -55,7 +55,7 @@ echo "Sequence Length: $SEQ_LEN"
 echo "Running model $MODEL with PCA Attention and top-r $TOPR and top-k $TOPK"
 
 run_cmd="srun -C gpu -N $NNODES -n $GPUS -c 32 --cpu-bind=cores --gpus-per-node=4 ./set_env_vars_slurm.sh\
-        python -u eval_ppl.py --sequence-length $SEQ_LEN --model-id $MODEL --model-type $MODEL_TYPE\
+        python -u evaluate_tasks.py --sequence-length $SEQ_LEN --model-id $MODEL --model-type $MODEL_TYPE\
         --use-pca-topk --top-r $TOPR --top-k $TOPK --dataset $TDATA --rotary-type $ROTARY\
         $WANDB_ARGS $AXONN_ARGS $EVAL_ARGS"
